list view is fully working
